==> notebooks/step_1.ipynb <==
### Step 1 - Retrieve and Extract Subreddit Data

At the time of writing the pushshift API was down and historical posts were not available.  As a backup, historical subreddit posts from academictorrents.com were retrieved.  Academictorrents.com maintains a repository of the pushshift data.  Data was retrieved from 2005-06 through 2022-12.  Details can be found here: https://academictorrents.com/details/c398a571976c78d346c325bd75c47b82edf6124e.

Data was retrieved in a compressed format.  The uncompressed format is an NDJSON file format.

!ls

#### Install Dependencies

!pip install openai

!pip install bertopic

!pip install bert-score

#### Import required libraries and read NDJSON file into a pandas dataframe.  Limit to required fields only.

import pandas as pd
import json 
import csv 
import time
import datetime
import os
import re

filename = 'talesfromcallcenters_submissions.ndjson' 

# Read the JSON file directly into a dataframe, selecting the desired columns
results = pd.read_json(filename, lines=True)[['id', 'title', 'selftext', 'author', 'score', 'num_comments', 'created_utc']]

#### Drop NaN and view the first 10 rows of the dataframe

results.dropna(axis=1, inplace=True)

total_submissions = num_rows = len(results.index)


results

#### Note the [deleted] value in author field.  Check to see if this occurs in the selftext field as well.  Selftext is the body of the submission.  Selftext value of [deleted] will not be useful for the anlaysis.



# Count the number of times "[deleted]" appears in the specified column
num_deleted = (results['selftext'] == '[deleted]').sum()

# Print the result
print(f'The word "[deleted]" appears {num_deleted} times in the selftext column')

#### Since the word [deleted] appears we remove those rows from the results in the next step.  Author [deleted], if any, are OK though since we still have the selftext to analyze.

results = results[~results['selftext'].isin(['[deleted]', '[removed]'])]

total_usable_submissions = num_rows = len(results.index)


#### Display the results to get the total row count and a view of the first five and last five rows

results

#### The created_utc field is showing time from epoch in seconds which is not useful.  This needs to be converted to a readable format and placed in a new field called created_date



results['created_date'] = pd.to_datetime(results['created_utc'], unit='s')

# Drop the epoch seconds column
results.drop('created_utc', axis=1, inplace=True)




#### Check the dataframe again to ensure that the new field has been populated correctly and the old field dropped

results

# extract earliest and latest dates
start_date = results['created_date'].min().date()
end_date = results['created_date'].max().date()

# print results
print("Start date: ", start_date)
print("End date: ", end_date)

#### Remove posts shorter than 100 words

def word_count(text):
    return len(text.split())

results = results[results['selftext'].apply(word_count) >= 100]

#### Remove posts that are greater than 874 tokens

from transformers import AutoTokenizer

# Initialize the tokenizer
tokenizer = AutoTokenizer.from_pretrained('google/pegasus-large')

# Define the token count threshold
max_tokens = 874

# Function to count tokens in a text
def count_tokens(text):
    return len(tokenizer.encode(text, truncation=False, max_length=1024))

# Apply the token count function to the 'selftext' column
results['token_count'] = results['selftext'].apply(count_tokens)

# Filter the dataframe
filtered_submissions = results[results['token_count'] <= max_tokens]

# Drop the 'token_count' column as it's not needed anymore
filtered_df = filtered_submissions.drop(columns=['token_count'])

#### Get number of rows

filtered_df

total_inscope_submissions = num_rows = len(filtered_df.index)


#### Clean up the text

def clean_text(text):
    # Remove emojis
    text = re.sub(r'[\U00010000-\U0010ffff]', '', text)
    # Remove hyperlinks
    text = re.sub(r'http\S+', '', text)

    # Remove unicode characters
    text = re.sub(r'[^\x00-\x7F]+', '', text)

    # Remove HTML codes
    text = text.replace("&amp;nbsp;", "")

    # Replace 'tl;dr' with 'in summary'
    new_text = text.replace("tl;dr", "in summary")

    return new_text

# Apply the function to the dataframe
filtered_df['selftext'] = filtered_df['selftext'].apply(clean_text)

#### Save work to this point by exporting the results dataframe to a CSV

filtered_df.to_csv('tfcc_submissions.csv', index=False)

#### Repeat above steps for comments.  In this case we will only retrieve parent_id (to associate the comment with an original submission), body, and score

##talesfromcallcenters_comments

filename = 'talesfromcallcenters_comments.ndjson'

# Read the JSON file directly into a dataframe, selecting the desired columns
results = pd.read_json(filename, lines=True)[['parent_id', 'body', 'score']]




#### Drop NaN and display the first few and last rows.

results.dropna(axis=1, inplace=True)

total_comments = num_rows = len(results.index)

results

#### Remove all comments with less than 50 words

def word_count(text):
    return len(text.split())

results = results[results['body'].apply(word_count) >= 50]


results

#### Remove comments > 200 tokens

from transformers import AutoTokenizer

# Initialize the tokenizer
tokenizer = AutoTokenizer.from_pretrained('google/pegasus-large')

# Define the token count threshold
max_tokens = 200

# Function to count tokens in a text
def count_tokens(text):
    return len(tokenizer.encode(text, truncation=False, max_length=1024))

# Apply the token count function to the 'selftext' column
results['token_count'] = results['body'].apply(count_tokens)

# Filter the dataframe
filtered_comments = results[results['token_count'] <= max_tokens]

# Drop the 'token_count' column as it's not needed anymore
filtered_comments_df = filtered_comments.drop(columns=['token_count'])

total_in_scope_comments = num_rows = len(filtered_comments_df.index)


filtered_comments_df

#### Clean up the text

filtered_comments_df['body'] = filtered_comments_df['body'].apply(clean_text)

#### Export the comments to a csv

filtered_comments_df.to_csv('tfcc_comments.csv', index=False)

#### Clean up.  Remove the NDJSON files from the working directory


os.remove('talesfromcallcenters_comments.ndjson')

os.remove('talesfromcallcenters_submissions.ndjson')


### Wrap up: We now have two CSV files of reddit submissions and posts ready for analysis

### This section creates a CSV file for the introduction section of the RMarkdown report

#### Create the subreddit meta data table

import openai
import json 
openai.api_key = ""

import json
import openai

error_string = "{\"summary\": \"none\"}"

def getOpenAI_Summary(prompt):
    response = openai.ChatCompletion.create(
        model="gpt-3.5-turbo",
        messages=[
            {"role": "user", "content": "" + prompt + ""}
        ]
    )

    initial_response = response['choices'][0]['message']['content']

    print(initial_response)

    return initial_response


#### Because Cohere and OpenAI training data included Reddit, it will be able to answer general questions about a specific subreddit.  Here we ask OpenAI to describe the subreddit

prompt = 'Describe the Tales from Call Centers (TFCC) subreddit in two lengthy paragraphs.  Include things like history, volume, and themes.  Write in an academic style:'

subreddit_description = getOpenAI_Summary(prompt)


subreddit_name = "Tales from Call Centers (TFCC)"
subreddit_url = "https://www.reddit.com/r/talesfromcallcenters/"
total_submissions
total_usable_submissions
total_inscope_submissions
total_comments = num_rows
total_in_scope_comments


#### Construct a string that described the number of submissions and comments that were in scope and the date range of the subreddit data

scope = "A total of " + format(total_submissions, ',') + " subreddit submissions and " + format(total_comments, ',') + " associated comments were extracted for the period of " + str(start_date) + " through " + str(end_date) + " from the " + subreddit_name + " subreddit (" + subreddit_url + "). Of the total submissions and comments, " + format(total_usable_submissions, ',') + " submissions and " + format(total_in_scope_comments, ',') + " comments were retained after cleanup (short text, null value removal)."


scope2 = " The following sections summarize each of the top 20 topics identified through topic modeling using the BERTopic library.  The sections are formatted as follows: 1) Topic number plus the BERTopic description 2) Themes identified from OpenAI 3) Sentiment analysis results (note that while all sentiment scores are negative, the comments sentiment are generally more positive than the submissions sentiment) 4) Summaries of Submissions (each paragraph represents between 25 - 50 original submissions that have been summarized twice using abstractive summarization) and 5) Summary of comments (a brief summarization of the top 100 comments related to the topic)."

scope = scope + scope2

description_intro = "The following report was generated completely using the OpenAI GPT 3.5 Turbo API through a series of summarization steps.  It is important to note that one of the common risks associated with abstractive summarization is hallucination, which is the introduction of content not completely relevant to the source text.  Abstractive summarization is not perfect and while there are methods to check the accuracy of an abstractive summary, it is not a guarantee of accuracy.  The intent of this report is to consolidate the submissions to a subreddit over an extended period of time and group those submissions into categories identified using NLP analysis techniques.  Should some of the topics be of interest, then a further review of the original subreddit posts is recommended.  Stephen Drew, 2 April 2023."

scope

# create a dictionary with the variables as values
data = {'scope': [scope], 'subreddit_description': [subreddit_description], 'subreddit_name': [subreddit_name], 'description_intro': [description_intro]}

# create a DataFrame from the dictionary
df = pd.DataFrame(data)

df

#### Export to a CSV

df.to_csv('subreddit_overview.csv', index=False)
