==> notebooks/step_5.ipynb <==
### Step 5. Topic Modeling of Submissions

#### Import required libraries and read submissions csv into a Pandas dataframe

import pandas as pd
from datasets import Dataset
import nltk, re
from bertopic import BERTopic
from bertopic.representation import MaximalMarginalRelevance
from sentence_transformers import SentenceTransformer


#### Import CSV file from steps 2 - 4

tfcc_data = pd.read_csv('tfcc_submissions_for_top_modeling.csv')


from nltk.corpus import stopwords
from nltk.stem import PorterStemmer
from nltk.tokenize import word_tokenize, sent_tokenize
from nltk.tokenize import RegexpTokenizer
from nltk.stem.wordnet import WordNetLemmatizer

from collections import Counter
from nltk.util import ngrams
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('omw-1.4')

#### Setup stop words and add additional words relevant to the subreddit (these can be found through the first one or two passes of topic modeling and may require some trial and error to identify)

Note: Remove work, time, day from stop words for first pass.  Add back in when running for -1 topic

stop_words = stopwords.words('english')
stop_words.extend(['call', 'customer', 'phone', 'company', 'caller', 'account', 'calls','center','x200b', 'work', 'time', 'day'])

print(len(stop_words))

#### Convert text to lower case, tokenize self text field, remove stop words, lemmatize tokens

data = tfcc_data

#Transforming all text to lower case
data['selftext'] = data['selftext'].astype(str).str.lower()
#data.head()

#Tokenizing selftext
tokenizer = RegexpTokenizer('\w+')
data['tokens'] = data['selftext'].apply(tokenizer.tokenize)
#data.head()

#Removning stop words using lambda
data['tokens'] = data['tokens'].apply(lambda x: [item for item in x if item not in stop_words])
#data.head()

#Converting to a string and keeping longer than 2 chars
data['tokens'] = data['tokens'].apply(lambda x: ' '.join([item for item in x if len(item) > 2]))
#data.head()

#Lemmatization of tokens
lemmatizer = WordNetLemmatizer()
data['lemmatized'] = data['tokens'].apply(lemmatizer.lemmatize)
data.head()

#### Convert pandas dataframe to huggingface dataset

tfcc_data = data.to_dict('list')
    
# Create a Hugging Face dataset from the dictionary
dataset = Dataset.from_dict(tfcc_data)

dataset

#### Create topic model with BERTopic.  This takes about 1 min on a laptop with 8 GB GPU, 32 GB RAM

While the documentation for BERTopic indicates stop word removal and tokenization may not be necessary, I found better results when including these steps

NOTE: when running the second time for the -1 topics, update diversity to 0.4

sentence_model = SentenceTransformer("all-MiniLM-L12-v2")
embeddings = sentence_model.encode(dataset["lemmatized"], show_progress_bar=False)
representation_model = MaximalMarginalRelevance(diversity=0.4)
model = BERTopic(embedding_model=sentence_model, representation_model=representation_model)
topics, probs = model.fit_transform(dataset['lemmatized'])

model.get_topic_info()

details = model.get_topic_info()

#### Update the dataframe to include the topic number for each submission

data['topic'] = topics

details

#### There are a high number of topic -1, which are basically submissions that fall into a general category.  We'll save these to a separate CSV to look at later

neg1_topic = data.loc[(data['topic'] == -1)]
neg1_topic.to_csv('tfcc_submissions_neg1_topics.csv', index=False)

#### Take a look at the top 20 topics

model.visualize_barchart(top_n_topics=20, n_words=5)


### Filter dataset to only the topics of interest for analysis.  In this case we'll keep top 20 topics which are 0 - 19.  Tokens and Lemmatized columns are no longer required and are dropped.

data = data.drop(['tokens', 'lemmatized'], axis=1)


data = data.loc[(data['topic'] >= 0) & (data['topic'] <= 19)]


data

### The submission counts for the topics

details.head(20)

#### Export the filtered dataset and the topic counts for the top 20 topics.

details.to_csv('tfcc_submissions_topic_counts.csv', index=False)

data.to_csv('tfcc_submissions_selected_topics.csv', index=False)

#### Run from top again on the -1 topics from first pass
