==> notebooks/step_10_1.ipynb <==
### Step 10.  Accuracy Measurement

This notebook measuring accuracy with ROUGE and is not included in the dissertation.  Not required for reproducing results.

#### Import required libraries, load submissions and comments with summary

import torch
import pandas as pd
from rouge import Rouge




def extract_nan_rows(df, name):
    nan_rows = df[df['summary'].isna()]
    nan_rows['source'] = name
    return nan_rows

# Load DataFrames
df = pd.read_csv('tfcc_submissions_top20_with_sentiment_including_comment_sentiment_and_summaries.csv')
comments_df = pd.read_csv('tfcc_top_comments_summarized.csv')
pegasus_df = pd.read_csv('tfcc_submissions_top20_pegasus_summaries.csv')
pegasus_comments_df = pd.read_csv('tfcc_top_comments_pegasus_summarized.csv')
cohere_df = pd.read_csv('tfcc_submissions_top20_cohere_summaries.csv')
cohere_comments_df = pd.read_csv('tfcc_top_comments_cohere_summarized.csv')

# Extract rows with NaN values in the 'summary' column
nan_rows_df = pd.DataFrame()
nan_rows_df = nan_rows_df.append(extract_nan_rows(df, "tfcc_submissions_top20_with_sentiment_including_comment_sentiment_and_summaries"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(comments_df, "tfcc_top_comments_summarized"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(pegasus_df, "tfcc_submissions_top20_pegasus_summaries"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(pegasus_comments_df, "tfcc_top_comments_pegasus_summarized"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(cohere_df, "tfcc_submissions_top20_cohere_summaries"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(cohere_comments_df, "tfcc_top_comments_cohere_summarized"))

# Remove rows with NaN values from the original DataFrames
df = df.dropna(subset=['summary'])
comments_df = comments_df.dropna(subset=['summary'])
pegasus_df = pegasus_df.dropna(subset=['summary'])
pegasus_comments_df = pegasus_comments_df.dropna(subset=['summary'])
cohere_df = cohere_df.dropna(subset=['summary'])
cohere_comments_df = cohere_comments_df.dropna(subset=['summary'])

# Print the DataFrame containing rows with NaN values in the 'summary' column
nan_rows_df.to_csv('summary_errors.csv', index=False)



nan_rows_df

### Measure rouge

# get the number of rows using the shape attribute
start_num_rows = df.shape[0]
pegasus_start_num_rows = pegasus_df.shape[0]
cohere_start_num_rows = cohere_df.shape[0]

print("OpenAI - Measuring accuracy on " +str(start_num_rows) + " rows.")
print("Pegasus - Measuring accuracy on " +str(pegasus_start_num_rows) + " rows.")
print("COhere - Measuring accuracy on " +str(cohere_start_num_rows) + " rows.")


# Initialize the Rouge object
rouge = Rouge()

# Function to calculate ROUGE scores
def calculate_rouge1_score(row):
    try:
        scores = rouge.get_scores(row['summary'], row['selftext'])
        return scores[0]['rouge-1']  # Return only the ROUGE-1 score
    except Exception as e:
        print(f"Error calculating ROUGE-1 score: {e}")
        return None

def extract_f1_score(row):
    return row['rouge1_scores']['f']



#### Calculate rouge for OpenAI submissions

# Apply the function to the dataframe
df['rouge1_scores'] = df.apply(calculate_rouge1_score, axis=1)

# Extract F1-score and store it in a new column called 'rouge1_f1_score'
df['rouge1_f1_score'] = df.apply(extract_f1_score, axis=1)

# Calculate the average ROUGE-1 F1-score
average_rouge1_f1 = df['rouge1_f1_score'].mean()

# Print the average ROUGE-1 F1-score
print(f"Average ROUGE-1 F1-score: {average_rouge1_f1}")




#### Calculate rouge for Pegasus submissions

# Apply the function to the dataframe
pegasus_df['rouge1_scores'] = pegasus_df.apply(calculate_rouge1_score, axis=1)

# Extract F1-score and store it in a new column called 'rouge1_f1_score'
pegasus_df['rouge1_f1_score'] = pegasus_df.apply(extract_f1_score, axis=1)

# Calculate the average ROUGE-1 F1-score
average_rouge1_f1 = pegasus_df['rouge1_f1_score'].mean()

# Print the average ROUGE-1 F1-score
print(f"Average ROUGE-1 F1-score: {average_rouge1_f1}")


#### Calculate rouge for Cohere submissions

# Apply the function to the dataframe
cohere_df['rouge1_scores'] = cohere_df.apply(calculate_rouge1_score, axis=1)

# Extract F1-score and store it in a new column called 'rouge1_f1_score'
cohere_df['rouge1_f1_score'] = cohere_df.apply(extract_f1_score, axis=1)

# Calculate the average ROUGE-1 F1-score
average_rouge1_f1 = cohere_df['rouge1_f1_score'].mean()

# Print the average ROUGE-1 F1-score
print(f"Average ROUGE-1 F1-score: {average_rouge1_f1}")
