==> notebooks/step_12.ipynb <==
### Step 12.  Measure Summaries of Summaries Accuracy for Submissions and Comments

#### Import required libraries, load submissions and comments with summary

import torch
import bert_score

import pandas as pd

def extract_nan_rows(df, name):
    nan_rows = df[df['new_summary'].isna()]
    nan_rows['source'] = name
    return nan_rows

# Load DataFrames
df = pd.read_csv('tfcc_top_20_summaries_of_summaries.csv')
comments_df = pd.read_csv('tfcc_comments_summaries_of_summaries.csv')
pegasus_df = pd.read_csv('tfcc_top_20_summaries_of_summaries_pegasus.csv')
pegasus_comments_df = pd.read_csv('tfcc_comments_summaries_of_summaries_pegasus.csv')
cohere_df = pd.read_csv('tfcc_top_20_summaries_of_summaries.csv')
cohere_comments_df = pd.read_csv('tfcc_comments_summaries_of_summaries_cohere.csv')

# Extract rows with NaN values in the 'summary' column
nan_rows_df = pd.DataFrame()
nan_rows_df = nan_rows_df.append(extract_nan_rows(df, "tfcc_top_20_summaries_of_summaries"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(comments_df, "tfcc_comments_summaries_of_summaries"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(pegasus_df, "tfcc_top_20_summaries_of_summaries_pegasus"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(pegasus_comments_df, "tfcc_comments_summaries_of_summaries_pegasus"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(cohere_df, "tfcc_top_20_summaries_of_summaries"))
nan_rows_df = nan_rows_df.append(extract_nan_rows(cohere_comments_df, "tfcc_comments_summaries_of_summaries_cohere"))

# Remove rows with NaN values from the original DataFrames
df = df.dropna(subset=['new_summary'])
comments_df = comments_df.dropna(subset=['new_summary'])
pegasus_df = pegasus_df.dropna(subset=['new_summary'])
pegasus_comments_df = pegasus_comments_df.dropna(subset=['new_summary'])
cohere_df = cohere_df.dropna(subset=['new_summary'])
cohere_comments_df = cohere_comments_df.dropna(subset=['new_summary'])

# Print the DataFrame containing rows with NaN values in the 'summary' column
nan_rows_df.to_csv('summary_errors.csv', index=False)



nan_rows_df

### Measure summary accuracy and drop all rows that are below 0.8 F1

# get the number of rows using the shape attribute
start_num_rows = df.shape[0]
pegasus_start_num_rows = pegasus_df.shape[0]
cohere_start_num_rows = cohere_df.shape[0]

print("OpenAI - Measuring accuracy on " +str(start_num_rows) + " rows.")
print("Pegasus - Measuring accuracy on " +str(pegasus_start_num_rows) + " rows.")
print("COhere - Measuring accuracy on " +str(cohere_start_num_rows) + " rows.")


#### Calculate accuracy for OpenAI submissions

import torch

selftext = df['summary']
summary = df['new_summary']
P, R, F1 = bert_score.score(selftext.tolist(), summary.tolist(), lang='en', verbose=True)
df['bert_f1'] = F1.tolist()

# Free up GPU memory
del P, R, F1
torch.cuda.empty_cache()




#### Calculate accuracy for Pegasus submissions

selftext = pegasus_df['summary']
summary = pegasus_df['new_summary']

# Compute the BERTScore for each pair of text
P, R, F1 = bert_score.score(selftext.tolist(), summary.tolist(), lang='en', verbose=True)
pegasus_df['bert_f1'] = F1.tolist()

# Free up GPU memory
del P, R, F1
torch.cuda.empty_cache()


#### Calculate accuracy for Cohere submissions

selftext = cohere_df['summary']
summary = cohere_df['new_summary']

# Compute the BERTScore for each pair of text
P, R, F1 = bert_score.score(selftext.tolist(), summary.tolist(), lang='en', verbose=True)
cohere_df['bert_f1'] = F1.tolist()

# Free up GPU memory
del P, R, F1
torch.cuda.empty_cache()


#### Display average accuracy score for OpenAI, Pegasus, and Cohere

from IPython.display import display, HTML


# Calculate the average F1 scores
df_avg_f1 = df['bert_f1'].mean()
pegasus_avg_f1 = pegasus_df['bert_f1'].mean()
cohere_avg_f1 = cohere_df['bert_f1'].mean()

# Create a dictionary to store the results
data = {
    'Dataframe': ['df', 'pegasus_df', 'cohere_df'],
    'Average F1': [df_avg_f1, pegasus_avg_f1, cohere_avg_f1]
}

# Create a new dataframe to display the results
results_df = pd.DataFrame(data)

# Display the results in a simple table
display(HTML(results_df.to_html(index=False)))


#### Print results for OpenAI test

df2 = df[df['bert_f1'] < 0.8] 
df = df[df['bert_f1'] >= 0.8]

end_num_rows = df.shape[0]

removed_num = df2.shape[0]

print(str(end_num_rows) + " rows were >= 0.8 F1 and " + str(removed_num) + " were below 0.8 F1 and have been removed.")


df.to_csv('tfcc_submissions_summaries_of_summaries_with_accuracy.csv', index=False)

#### Print results for Pegasus test

pegasus_df2 = pegasus_df[pegasus_df['bert_f1'] < 0.8] 
pegasus_df = pegasus_df[pegasus_df['bert_f1'] >= 0.8]

end_num_rows = pegasus_df.shape[0]

removed_num = pegasus_df2.shape[0]

print(str(end_num_rows) + " rows were >= 0.8 F1 and " + str(removed_num) + " were below 0.8 F1 and have been removed.")

pegasus_df.to_csv('tfcc_submissions_summaries_of_summaries_pegasus_with_accuracy.csv', index=False)

#### Print results for Cohere test

temp_df2 = cohere_df[cohere_df['bert_f1'] < 0.8] 
temp_df = cohere_df[cohere_df['bert_f1'] >= 0.8]

end_num_rows = temp_df.shape[0]

removed_num = temp_df2.shape[0]

print(str(end_num_rows) + " rows were >= 0.8 F1 and " + str(removed_num) + " were below 0.8 F1.")

cohere_df.to_csv('tfcc_submissions_summaries_of_summaries_cohere_with_accuracy.csv', index=False)

### Repeat accuracy measurement process for comments.  

# get the number of rows using the shape attribute
start_num_rows = comments_df.shape[0]
print("OpenAI Measuring accuracy on " +str(start_num_rows) + " rows.")

pegasus_start_num_rows = pegasus_comments_df.shape[0]
print("Pegasus Measuring accuracy on " +str(pegasus_start_num_rows) + " rows.")

cohere_start_num_rows = cohere_comments_df.shape[0]
print("Cohere Measuring accuracy on " +str(cohere_start_num_rows) + " rows.")

#### Measure OpenAI comments summaries accuracy

# Select the columns to compare
selftext = comments_df['summary']
summary = comments_df['new_summary']

# Compute the BERTScore for each pair of text
P, R, F1 = bert_score.score(selftext.tolist(), summary.tolist(), lang='en', verbose=True)
comments_df['bert_f1'] = F1.tolist()

# Free up GPU memory
del P, R, F1
torch.cuda.empty_cache()

#### Measure Pegasus comments summaries accuracy

# Select the columns to compare
selftext = pegasus_comments_df['summary']
summary = pegasus_comments_df['new_summary']

# Compute the BERTScore for each pair of text
P, R, F1 = bert_score.score(selftext.tolist(), summary.tolist(), lang='en', verbose=True)
pegasus_comments_df['bert_f1'] = F1.tolist()

# Free up GPU memory
del P, R, F1
torch.cuda.empty_cache()

#### Measure Cohere comments summaries accuracy

cohere_comments_df

# Select the columns to compare
selftext = cohere_comments_df['concatenated_summary']
summary = cohere_comments_df['new_summary']

# Compute the BERTScore for each pair of text
P, R, F1 = bert_score.score(selftext.tolist(), summary.tolist(), lang='en', verbose=True)
cohere_comments_df['bert_f1'] = F1.tolist()

# Free up GPU memory
del P, R, F1
torch.cuda.empty_cache()

#### Display average accuracy score for OpenAI, Pegasus, and Cohere

# Calculate the average F1 scores
df_avg_f1 = comments_df['bert_f1'].mean()
pegasus_avg_f1 = pegasus_comments_df['bert_f1'].mean()
cohere_avg_f1 = cohere_comments_df['bert_f1'].mean()

# Create a dictionary to store the results
data = {
    'Dataframe': ['df', 'pegasus_df', 'cohere_df'],
    'Average F1': [df_avg_f1, pegasus_avg_f1, cohere_avg_f1]
}

# Create a new dataframe to display the results
results_df = pd.DataFrame(data)

# Display the results in a simple table
display(HTML(results_df.to_html(index=False)))

#### Print OpenAI results

temp_df2 = comments_df[comments_df['bert_f1'] < 0.8] 
temp_df = comments_df[comments_df['bert_f1'] >= 0.8]

end_num_rows = temp_df.shape[0]

removed_num = temp_df2.shape[0]

print(str(end_num_rows) + " rows were >= 0.8 F1 and " + str(removed_num) + " were below 0.8 F1.")


#### Print Pegasus results

temp_df2 = pegasus_comments_df[pegasus_comments_df['bert_f1'] < 0.8] 
temp_df = pegasus_comments_df[pegasus_comments_df['bert_f1'] >= 0.8]

end_num_rows = temp_df.shape[0]

removed_num = temp_df2.shape[0]

print(str(end_num_rows) + " rows were >= 0.8 F1 and " + str(removed_num) + " were below 0.8 F1.")


#### Print Cohere results

temp_df2 = cohere_comments_df[cohere_comments_df['bert_f1'] < 0.8] 
temp_df = cohere_comments_df[cohere_comments_df['bert_f1'] >= 0.8]

end_num_rows = temp_df.shape[0]

removed_num = temp_df2.shape[0]

print(str(end_num_rows) + " rows were >= 0.8 F1 and " + str(removed_num) + " were below 0.8 F1.")

#### Double check that we still have comments for every top.  Count should = 20

OpenAI:

unique_count = comments_df['topic'].nunique()

print(unique_count)

Pegasus:

unique_count = pegasus_comments_df['topic'].nunique()

print(unique_count)

Cohere:

unique_count = cohere_comments_df['topic'].nunique()

print(unique_count)

#### Export comments with accuracy to CSV

comments_df.to_csv('tfcc_comments_summaries_of_summaries_with_accuracy.csv', index=False)
pegasus_comments_df.to_csv('tfcc_pegasus_comments_summaries_of_summaries_with_accuracy.csv', index=False)
cohere_comments_df.to_csv('tfcc_cohere_comments_summaries_of_summaries_with_accuracy.csv', index=False)

#### Generate accuracy report

import pandas as pd

openai_df = pd.read_csv('tfcc_submissions_summaries_of_summaries_with_accuracy.csv')
pegasus_df = pd.read_csv('tfcc_submissions_summaries_of_summaries_pegasus_with_accuracy.csv')
cohere_df = pd.read_csv('tfcc_submissions_summaries_of_summaries_cohere_with_accuracy.csv')

openai_comments_df = pd.read_csv('tfcc_comments_summaries_of_summaries_with_accuracy.csv')
pegasus_comments_df = pd.read_csv('tfcc_pegasus_comments_summaries_of_summaries_with_accuracy.csv')
cohere_comments_df = pd.read_csv('tfcc_cohere_comments_summaries_of_summaries_with_accuracy.csv')


topics = pd.read_csv('tfcc_top20_topics_with_sentiment_and_comments_sentiment.csv')

topics.rename(columns={'Topic': 'topic'}, inplace=True)


import pandas as pd

# Group by 'topic' and calculate the mean F1 score for each model
openai_average_f1 = openai_df.groupby('topic')['bert_f1'].mean().reset_index().rename(columns={'bert_f1': 'openai_bert_f1'})
pegasus_average_f1 = pegasus_df.groupby('topic')['bert_f1'].mean().reset_index().rename(columns={'bert_f1': 'pegasus_bert_f1'})
cohere_average_f1 = cohere_df.groupby('topic')['bert_f1'].mean().reset_index().rename(columns={'bert_f1': 'cohere_bert_f1'})

# Merge the topics dataframe with the average F1 scores dataframes on the 'topic' column
topics_with_f1 = topics.merge(openai_average_f1, on='topic', how='left')
topics_with_f1 = topics_with_f1.merge(pegasus_average_f1, on='topic', how='left')
topics_with_f1 = topics_with_f1.merge(cohere_average_f1, on='topic', how='left')


topics_with_f1

topics_with_f1 = topics_with_f1.drop(columns=['Count', 'avg_pos_sentiment', 'avg_neg_sentiment', 'avg_comments_pos_sentiment', 'avg_comments_neg_sentiment'])


topics_with_f1

# Group by 'topic' and calculate the mean F1 score for each model
openai_average_f1 = openai_comments_df.groupby('topic')['bert_f1'].mean().reset_index().rename(columns={'bert_f1': 'openai_bert_f1'})
pegasus_average_f1 = pegasus_comments_df.groupby('topic')['bert_f1'].mean().reset_index().rename(columns={'bert_f1': 'pegasus_bert_f1'})
cohere_average_f1 = cohere_comments_df.groupby('topic')['bert_f1'].mean().reset_index().rename(columns={'bert_f1': 'cohere_bert_f1'})

# Merge the topics dataframe with the average F1 scores dataframes on the 'topic' column
comments_topics_with_f1 = topics.merge(openai_average_f1, on='topic', how='left')
comments_topics_with_f1 = comments_topics_with_f1.merge(pegasus_average_f1, on='topic', how='left')
comments_topics_with_f1 = comments_topics_with_f1.merge(cohere_average_f1, on='topic', how='left')

comments_topics_with_f1 = comments_topics_with_f1.drop(columns=['Count','avg_pos_sentiment', 'avg_neg_sentiment', 'avg_comments_pos_sentiment', 'avg_comments_neg_sentiment'])


comments_topics_with_f1
